+++ benchmark/tpch/CMakeLists.txt
@@ -1,5 +1,7 @@
 include_directories(../../third_party/dbgen/include)
 include_directories(../../third_party/sqlite/include)
+include_directories(../../extension/parquet/include)
+
 add_library(duckdb_benchmark_tpch OBJECT
             lineitem_aggregate.cpp
             lineitem_parquet.cpp
